requirements.txt
pathlib==1.0.1
PyYAML==6.0.1
setuptools==70.3.0
